+++ src/methods.ts
@@ -1,0 +1,4 @@
+
+const Methods = [ "GET", "PATCH", "PUT", "POST", "DELETE" ] as const;
+
+export { Methods };